requirements.txt
supabase==2.3.1
httpx==0.25.0
websockets==12.0
python-dotenv==1.0.0
PySide6==6.7.0